Update dockyards cluster controller to await deployment conditions
Currently a cluster will be marked as ready as soon as the cluster and
control plane is ready. This is somewhat of a false positive because the
cluster is not really ready to use.

This update the ready condition to wait for all cluster component
deployments to become ready as well.