Update dockyards cluster controller to not use cluster condition reason
Currently the dockyards cluster controller uses the cluster ready
condition reason directly as the reason on dockyards cluster resources.
This causes issues when the string in the reason is a valid cluster-api
condition but not a valid meta condition.

This updates the dockyards cluster controller to put the reason in the
dockyards cluster message and instead use a generic not ready reason.