Update cluster controller to only use fallbacks when reason is invalid
Currently the dockyards cluster controller always puts the cluster-api
reason in the message of the meta condition. This causes the ready
condition of the dockyards cluster to be very uninformative and static.

This updates the dockyards cluster controller to only put the
cluster-api condition reason in the message when the reason is an
invalid meta reason. In most conditions the reason can be used making
the dockyards cluster condition more informative.