Add dockyards node controller
Currently the machine controller creates and updates dockyards nodes
related to the machines. This pattern does not follow the standard of
having one resource type to handle and makes the controller
unnecessarily complex.

This adds a separate controller for dockyards nodes. The conditions of
the dockyards nodes are synced with their owner machines.

DY-425 #done.