Add dockyards node pool controller
Currently dockyards node pools will be deleted before all dockyards
nodes are removed. This makes it possible for dockyards nodes to be
orphaned but invisible to users.

This adds a dockyards node pool controller to ensure that dockyards node
pools remains as long as their are any dockyards nodes related to it.